+++ backend/src/main/java/org/example/gradingcenter/service/impl/ParentServiceImpl.java
@@ -6,23 +6,20 @@ import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 import org.example.gradingcenter.configuration.ModelMapperConfig;
 import org.example.gradingcenter.data.dto.users.ParentDto;
-import org.example.gradingcenter.data.entity.Grade;
 import org.example.gradingcenter.data.entity.Role;
-import org.example.gradingcenter.data.entity.School;
 import org.example.gradingcenter.data.entity.enums.Roles;
+import org.example.gradingcenter.data.entity.users.Headmaster;
 import org.example.gradingcenter.data.entity.users.Parent;
 import org.example.gradingcenter.data.entity.users.Student;
-import org.example.gradingcenter.data.entity.users.Teacher;
 import org.example.gradingcenter.data.entity.users.User;
 import org.example.gradingcenter.data.repository.ParentRepository;
 import org.example.gradingcenter.data.repository.StudentRepository;
 import org.example.gradingcenter.data.repository.UserRepository;
-import org.example.gradingcenter.data.repository.specification.StudentSpecification;
+import org.example.gradingcenter.exceptions.AuthenticationFailureException;
+import org.example.gradingcenter.exceptions.AuthorizationFailureException;
 import org.example.gradingcenter.exceptions.DuplicateEntityException;
 import org.example.gradingcenter.exceptions.EntityNotFoundException;
-import org.example.gradingcenter.service.ParentService;
-import org.example.gradingcenter.service.RoleService;
-import org.example.gradingcenter.service.UserService;
+import org.example.gradingcenter.service.*;
 import org.example.gradingcenter.util.MapperUtil;
 import org.springframework.data.jpa.domain.Specification;
 import org.springframework.security.access.prepost.PreAuthorize;
@@ -32,7 +29,6 @@ import java.util.HashSet;
 import java.util.List;
 import java.util.Optional;
 
-import static org.example.gradingcenter.util.DataUtil.fetchObjectFromDb;
 import static org.example.gradingcenter.util.MapperUtil.entityToDto;
 import static org.example.gradingcenter.util.MapperUtil.mapList;
 
@@ -48,6 +44,7 @@ public class ParentServiceImpl implements ParentService {
     private final RoleService roleService;
     private final StudentRepository studentRepository;
     private final ModelMapperConfig mapperConfig;
+    private final AuthorizationService authService;
 
     @Override
     public List<ParentDto> getParents() {
@@ -82,7 +79,6 @@ public class ParentServiceImpl implements ParentService {
     }
 
     @Override
-    //@PreAuthorize("hasAuthority('ROLE_ADMIN')")
     @Transactional
     public ParentDto createParent(Long userId) {
         Optional<Parent> existingParent = parentRepository.findById(userId);
@@ -109,6 +105,9 @@ public class ParentServiceImpl implements ParentService {
     @Override
     @PreAuthorize("hasAnyAuthority('ROLE_ADMIN', 'ROLE_PARENT')")
     public ParentDto updateParent(ParentDto parentDto, long id) throws EntityNotFoundException {
+        if (!authService.hasAnyRole(Roles.ROLE_HEADMASTER, Roles.ROLE_ADMIN) && authService.getLoggedInUser().getId() != id) {
+            throw new AuthorizationFailureException(Parent.class, "update");
+        }
         Parent parent = this.parentRepository.findById(id)
                 .map((parentToUpdate) -> {
                     parentToUpdate.setFirstName(parentDto.getFirstName());
@@ -120,7 +119,11 @@ public class ParentServiceImpl implements ParentService {
     }
 
     @Override
+    @PreAuthorize("hasAnyAuthority('ROLE_ADMIN', 'ROLE_PARENT')")
     public void addChildToParent(Long parentId, String childEgn) {
+        if (!authService.hasAnyRole(Roles.ROLE_ADMIN) && authService.getLoggedInUser().getId() != parentId) {
+            throw new AuthorizationFailureException(Parent.class, "update");
+        }
         Parent parent = fetchParent(parentId);
         Student student = studentRepository
                 .findByEgn(childEgn.trim())
@@ -130,7 +133,11 @@ public class ParentServiceImpl implements ParentService {
     }
 
     @Override
+    @PreAuthorize("hasAnyAuthority('ROLE_ADMIN', 'ROLE_PARENT')")
     public void removeChildFromParent(Long parentId, Long childId) {
+        if (!authService.hasAnyRole(Roles.ROLE_ADMIN) && authService.getLoggedInUser().getId() != parentId) {
+            throw new AuthorizationFailureException(Parent.class, "update");
+        }
         Parent parent = fetchParent(parentId);
         Student student = studentRepository
                 .findById(childId)
@@ -140,7 +147,11 @@ public class ParentServiceImpl implements ParentService {
     }
 
     @Override
+    @PreAuthorize("hasAnyAuthority('ROLE_ADMIN', 'ROLE_PARENT')")
     public void deleteParent(long id) {
+        if (authService.getLoggedInUser().getId() != id && !authService.hasAnyRole(Roles.ROLE_ADMIN)) {
+            throw new AuthorizationFailureException(Headmaster.class, "delete");
+        }
         parentRepository.deleteById(id);
     }
